third_party/blink/web_tests/external/wpt/fetch/corb/response_block.tentative.sub.https-expected.txt

This is a testharness.js-based test.
FAIL ORB: Expect error response. assert_equals: expected "script errored" but got "script loaded"
FAIL !CORB: CORB would have expected an empty response. assert_not_equals: got disallowed value "script loaded"
Harness: the test ran to completion.
